// example/src/types.ts
export type Board = {
  type: string;
  id: number;
  title: string;
  content: string;
  writerMemberRole: string;
  videoUrl: string | null;
  videoMetadata: string | null;
  contentType: string;
  imageList: string[];
  profileImageList: string[];
  writerImage: string;
  isPinned: boolean;
  isCreator: boolean;
  isLiked: boolean;
  isCreatorCommented: boolean;
  commentedCreatorOwnerImage: string | null;
  highlightedBy: string | null;
  highlighterProfileImage: string | null;
  category: string;
  isCreatorOwner: boolean;
  ownerSpaceName: string;
  ownerMemberRole: string;
  writerSpaceName: string;
  writerId: number;
  ownerId: number;
  likeCount: number;
  replyCount: number;
  viewCount: number;
  writerHandle: string;
  isMultiCategoryOwner: boolean;
  isAllPin: boolean;
  isCategoryPin: boolean;
  ogTags: string[];
  startAt: string | null;
  endAt: string | null;
  calendarType: string | null;
  calendarTypeColor: string | null;
  isAllDay: boolean;
  isPopular: string | null;
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
  categoryChangeable: boolean;
};

export type RootStackParamList = {
  ListScreen: undefined;
  DetailScreen: { item: Board };
};